03 - Gauge/002 - Agentic AI/0002 - AI Agents/4 - Agentic AI Recipe Agent/requirement.txt
phidata
groq
duckduckgo-search